refactors user subreducer, in progress working on adding all users view when logged in as admin
