+++ client/routes.js
@@ -6,7 +6,7 @@ import {LogIn, SignUp, Home, MyUserAccount, EditMyAccount} from './components'
 import {getMe} from './store'
 import AllProducts from './components/allproducts'
 import SingleProduct from './components/singleProduct'
-import allUsers from './components/allUsers'
+import AllUsers from './components/allUsers'
 import MyCart from './components/MyCart'
 
 /**
@@ -19,6 +19,8 @@ class Routes extends Component {
 
   render() {
     const {isLoggedIn} = this.props
+    const {isAdmin} = this.props
+    console.log(this.props, 'what is props?')
 
     return (
       <Switch>
@@ -35,6 +37,12 @@ class Routes extends Component {
             <Route exact path="/me" component={MyUserAccount} />
             <Route path="/me/edit" component={EditMyAccount} />
             <Route path="/mycart" component={MyCart} />
+            <Route path="/me/admin" component={AllUsers} />
+            {/* {isAdmin && (
+            <Switch> 
+              <Route path="/me/admin" component={allUsers} />
+            </Switch>
+              )} */}
           </Switch>
         )}
         {/* Displays our Login component as a fallback */}
@@ -51,7 +59,8 @@ const mapState = state => {
   return {
     // Being 'logged in' for our purposes will be defined has having a state.user that has a truthy id.
     // Otherwise, state.user will be an empty object, and state.user.id will be falsey
-    isLoggedIn: !!state.users.selected.id
+    isLoggedIn: !!state.users.selected.id,
+    isAdmin: state.users.isAdmin
   }
 }
 
@@ -72,5 +81,6 @@ export default withRouter(connect(mapState, mapDispatch)(Routes))
  */
 Routes.propTypes = {
   loadInitialData: PropTypes.func.isRequired,
-  isLoggedIn: PropTypes.bool.isRequired
+  isLoggedIn: PropTypes.bool.isRequired,
+  isAdmin: PropTypes.bool.isRequired
 }
